scripts/lucas/s2ag/requirements.txt
pandas>=1.5.0
smashed[remote]>=0.18.0
springs>=1.11.0
transformers>=4.20.0
s3fs>=2021.10.0
fsspec>=2021.10.0
pycld3
